// src/shared/sdk/chronos/models/Availability.ts
/**
 * Chronos API
 * API documentation for the Chronos application
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class Availability {
    'createdAt': Date;
    'dayOfWeek'?: number | null;
    'endTime': string;
    'id': number;
    'isRecurring': boolean;
    'specificDate'?: string | null;
    'staffId': number;
    'startTime': string;
    'tenantId': number;
    'updatedAt': Date;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "createdAt",
            "baseName": "created_at",
            "type": "Date",
            "format": "date-time"
        },
        {
            "name": "dayOfWeek",
            "baseName": "day_of_week",
            "type": "number",
            "format": "int32"
        },
        {
            "name": "endTime",
            "baseName": "end_time",
            "type": "string",
            "format": ""
        },
        {
            "name": "id",
            "baseName": "id",
            "type": "number",
            "format": "int32"
        },
        {
            "name": "isRecurring",
            "baseName": "is_recurring",
            "type": "boolean",
            "format": ""
        },
        {
            "name": "specificDate",
            "baseName": "specific_date",
            "type": "string",
            "format": "date"
        },
        {
            "name": "staffId",
            "baseName": "staff_id",
            "type": "number",
            "format": "int32"
        },
        {
            "name": "startTime",
            "baseName": "start_time",
            "type": "string",
            "format": ""
        },
        {
            "name": "tenantId",
            "baseName": "tenant_id",
            "type": "number",
            "format": "int32"
        },
        {
            "name": "updatedAt",
            "baseName": "updated_at",
            "type": "Date",
            "format": "date-time"
        }    ];

    static getAttributeTypeMap() {
        return Availability.attributeTypeMap;
    }

    public constructor() {
    }
}